fix: move hook to lib file
